Move raster_metadata into raster_boundary2shape for cutting of asf_time_series
