+++ hyp3lib/depreciated/asf_time_series.py
@@ -1,4 +1,3 @@
-import os
 from datetime import datetime, timedelta
 
 import netCDF4 as nc
@@ -10,7 +9,7 @@ from scipy.interpolate import interp1d
 from statsmodels.tsa.seasonal import seasonal_decompose
 
 from hyp3lib import GeometryError
-from hyp3lib.depreciated.asf_geometry import geometry_proj2geo, raster_meta
+from hyp3lib.depreciated.asf_geometry import geometry_proj2geo
 
 tolerance = 0.00005
 
@@ -245,68 +244,6 @@ def vector_meta(vectorFile):
     features.append(value)
 
   return (fields, proj, extent, features)
-
-
-def raster_metadata(input):
-
-  # Set up shapefile attributes
-  fields = []
-  field = {}
-  values = []
-  field['name'] = 'granule'
-  field['type'] = ogr.OFTString
-  field['width'] = 254
-  fields.append(field)
-  field = {}
-  field['name'] = 'epsg'
-  field['type'] = ogr.OFTInteger
-  fields.append(field)
-  field = {}
-  field['name'] = 'originX'
-  field['type'] = ogr.OFTReal
-  fields.append(field)
-  field = {}
-  field['name'] = 'originY'
-  field['type'] = ogr.OFTReal
-  fields.append(field)
-  field = {}
-  field['name'] = 'pixSize'
-  field['type'] = ogr.OFTReal
-  fields.append(field)
-  field = {}
-  field['name'] = 'cols'
-  field['type'] = ogr.OFTInteger
-  fields.append(field)
-  field = {}
-  field['name'] = 'rows'
-  field['type'] = ogr.OFTInteger
-  fields.append(field)
-  field = {}
-  field['name'] = 'pixel'
-  field['type'] = ogr.OFTString
-  field['width'] = 8
-  fields.append(field)
-
-  # Extract other raster image metadata
-  (outSpatialRef, outGt, outShape, outPixel) = raster_meta(input)
-  if outSpatialRef.GetAttrValue('AUTHORITY', 0) == 'EPSG':
-    epsg = int(outSpatialRef.GetAttrValue('AUTHORITY', 1))
-
-  # Add granule name and geometry
-  base = os.path.basename(input)
-  granule = os.path.splitext(base)[0]
-  value = {}
-  value['granule'] = granule
-  value['epsg'] = epsg
-  value['originX'] = outGt[0]
-  value['originY'] = outGt[3]
-  value['pixSize'] = outGt[1]
-  value['cols'] = outShape[1]
-  value['rows'] = outShape[0]
-  value['pixel'] = outPixel
-  values.append(value)
-
-  return (fields, values, outSpatialRef)
 
 
 def netcdf2boundary_mask(ncFile, geographic):
